% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-fun.R
\name{knit_exercise}
\alias{knit_exercise}
\title{Custom Knit function for RStudio}
\usage{
knit_exercise(inputFile, encoding = "UTF-8")
}
\arguments{
\item{inputFile}{The name of the R Markdown file to be rendered.}

\item{encoding}{As per \code{rmarkdown::render_site()} this is ignored.
The encoding is always assumed to be UTF-8.}
}
\description{
This function adds \code{-questions}/\code{-solutions} to the end of the filename before the file extension
of the output document from your R Markdown document when rendered by clicking the
Knit button in RStudio and when you have set \code{solutions: FALSE}/\code{solutions: TRUE} in the \verb{params:} field
of your YAML header.
}
\details{
This function is used with the \verb{knit:} field of the YAML header of an R Markdown file.
For example, if you want pdf output the YAML header would be as follows.

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{---
title: My Title
params:
  solutions: TRUE
output: pdf_document
knit: knitexercise::knit_exercise
---
}\if{html}{\out{</div>}}

This function also sets the global chunk option \code{include} to the value specified by
\verb{solutions:}, i.e. set \code{solutions: FALSE} for your questions document and
\code{solutions: TRUE} for your solutions document.
}
\examples{
\dontrun{
# To launch the R Markdown template run
rmarkdown::draft(
  'exercise.Rmd',
   template = 'exercise_template',
   package = 'knitexercise')
}

}
\seealso{
This function is modified from the following example by Yihui Xie
\url{https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html}.
}
