% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitexercise-package.r
\docType{package}
\name{knitexercise-package}
\alias{knitexercise-package}
\title{knitexercise: Knit exercise questions and solutions from the same R Markdown document}
\description{
This package provides a custom knit function, \code{knit_exercise}, which sets the global
chunk option \code{include}. This is controlled by setting the \code{solutions} parameter in
the YAML header of the R Markdown document.
}
\details{
Hence to generate your output document with questions set \code{solutions: FALSE}.
For your output document with your solutions set \code{solutions: TRUE}.

Your YAML header should look something like the following.

\if{html}{\out{<div class="sourceCode">}}\preformatted{---
title: My Title
author: My Name
params:
  solutions: FALSE
knit: knitexercise::knit_exercise
output: html_document
---
}\if{html}{\out{</div>}}

For your solutions output document \code{-solutions} is added to the filename before
the file extension.

There is a template R Markdown document which you can open through the:
\emph{File | New File | R Markdown... | From Template} menu or with the following code.

\if{html}{\out{<div class="sourceCode">}}\preformatted{rmarkdown::draft("my_exercise.Rmd", template = "exercise_template", package = "knitexercise")
}\if{html}{\out{</div>}}
}
