% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitexercise-package.r
\docType{package}
\name{knitexercise-package}
\alias{knitexercise-package}
\title{knitexercise: knit exercise questions and solutions from the same R Markdown document}
\description{
By setting the \code{solutions} parameter in the YAML header of your R Markdown document
this package provides a custom knit function which sets the global chunk option
\code{include} from that value (i.e \code{FALSE} for your questions document and \code{TRUE}
for your solutions document).
}
\details{
And you need to specify this function in the \verb{knit:} field of your YAML header.
So your YAML header should look something like the following.\preformatted{---
title: My Title
params:
  solutions: FALSE
knit: knitexercise::knit_exercise
output: html_document
---
}

For your solutions document \code{-solutions} is added to the output document filename
before the file extension.

There is a template R Markdown document which you can launch through the
File | New File | R Markdown... | From Template menu or with the following code.\preformatted{rmarkdown::draft("my_exercise.Rmd", template = "exercise_template", package = "knitexercise")
}
}
